requirements.txt
Flask==3.0.3
Flask-CORS==4.0.0
firebase-admin==6.2.0
pytest==7.4.3
python-dotenv==1.0.0
requests
google-genai